More init test and code coverage
